+++ HISTORY.md
@@ -9,6 +9,7 @@ v1.10.0 ()
 * Fixed timestamp issued caused by fix_time
 * Fixed websocket asserts on closed fd
 * Support more information on ubus
+* Support enable data cache from cloud
 
 
 v1.9.0 (2020-09-07)